ABSCamera: fix compiler warning
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@16571 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/ABS/include/stringtools.h
@@ -300,7 +300,7 @@ namespace str
   {
     std::wstring strGuid;
     strGuid.resize( 40, 0 );
-    StringFromGUID2( guid, (LPOLESTR)strGuid.c_str(), strGuid.size() );
+    StringFromGUID2( guid, (LPOLESTR)strGuid.c_str(), (int) strGuid.size() );
     return strGuid;
   }
